// src/app/demo/api/meter.model.ts
export interface Meter {
    unique_key?: string;
    name: string;
    dev_eui: string;
    serial: string;
    diameter: string;
    created_at?: string;
    type_communication: string;
    state: boolean;
    last_reading: number;
    model: string;
    class_support: string;
    battery_percentage: string;
    last_comunication: string | null;
    valve_state: string;
    latitude: number;
    longitude: number;
    company_unique_key: string;
    type_product: string;
    sub_red: string;
    app_skey: string;
    app_key: string;
    nwk_skey: string;
    dev_addr: string;
    direction_gprs: string;
    authentication_type: string;
    eui_aplication: string;
    imei: String;
    number_serial: String;
    gateway_unique_key: string;
    region: string;
}